fix(replicants): fix assignment sometimes being applied out-of-order (#492)
Fixes #489

I chose to fix this problem by no longer treating assignment separately from operations. Now, assignment is simply a type of operation called overwrite. This not only fixes this bug by ensuring that assignment uses the same queue as all other operations, but it simplifies the mental model of the code by making all actions that can be taken against a Replicant be an operation, instead of some being operations and some being assignment, each of which had different behavior and semantics.

BREAKING CHANGE: The order in which clients receive Replicant operations has changed. It is now more deterministic and correct, but it's possible that there is code in the wild which relied on the old, buggy behavior. Also, this means a new operation type has been added, so any Replicant change handlers which use the third operations argument may need to be updated to account for both the changed order of operations, and the existence of a new overwrite operation.